texture_convert/requirements.txt
layeredimage==2024.3
pillow==10.4.0